Rewrite form template with updated Appuniversum
Also using a new changeModel function. Using @value is deprecated, so we
need to listen for value changes ourselves and react to them. Instead of
writing all these setter function, this one setter function sets all the
properties on the model by giving it a value and the property name.