Retry accessing the original route after login
When the user tries to access a route when not logged in, the transition will be stored and retried by ESA. However, this only works as long as the user doesn't reload the page. The new ACM/IDM login setup requires a redirect, which causes the transition data to be lost. To work around this we store the url of the route in a cookie, which ESA also checks after login (for FastBoot support). As a result, the transition will now be tried again, even after a page reload.

Based on the code in the frontend-worship-decisions repo.