Add a workaround for a false-positive 403 request
Some endpoints in the dispatcher have an extra `mu-auth-allowed-groups` check, which is only updated by mu-auth when a request is handled.

Since we do the 403 check before doing any request, the complaint-form lookup fails and users see an error page.

By doing an extra `/accounts` we bypass the problem because that call should update the `mu-auth-allowed-groups` value before we try the actual API calls.